requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
google-ai-generativelanguage==0.4.0
google-api-core==2.17.1
google-auth==2.28.1
google-generativeai==0.3.2
googleapis-common-protos==1.62.0
grpcio==1.62.0
grpcio-status==1.62.0
idna==3.6
importlib-metadata==7.0.1
ipython==8.18.1
jedi==0.19.1
jsonpatch==1.33
jsonpointer==2.4
langchain-core==0.1.26
langchain-google-genai==0.0.9
langsmith==0.1.8
Markdown==3.5.2
matplotlib-inline==0.1.6
orjson==3.9.15
packaging==23.2
parso==0.8.3
pexpect==4.9.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.6.2
pydantic-core==2.16.3
pygments==2.17.2
pypdf==4.0.2
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
tenacity==8.2.3
tqdm==4.66.2
traitlets==5.14.1
typing-extensions==4.10.0
urllib3==2.2.1
wcwidth==0.2.13
zipp==3.17.0
